Rename CopyScheduler, mark Deprecated CopyScheduler as LegacyCopyScheduler.
