+++ smart-server/src/test/java/org/smartdata/server/TestLegacyCopyScheduler.java
@@ -30,14 +30,14 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.smartdata.model.FileDiff;
 import org.smartdata.model.FileDiffType;
-import org.smartdata.server.engine.CopyScheduler;
+import org.smartdata.server.engine.LegacyCopyScheduler;
 import org.smartdata.server.engine.CopyTargetTask;
 
 import java.io.IOException;
 import java.util.List;
 
 @Deprecated
-public class TestCopySchedular extends MiniSmartClusterHarness {
+public class TestLegacyCopyScheduler extends MiniSmartClusterHarness {
 
   @Rule
   public ExpectedException thrown = ExpectedException.none();
@@ -51,20 +51,23 @@ public class TestCopySchedular extends MiniSmartClusterHarness {
     // Create and write
     fileDiff.setParameters("");
     String cmd =
-        CopyScheduler.cmdParsing(fileDiff, "/root/", "/localhost:3306/backup/");
+        LegacyCopyScheduler
+            .cmdParsing(fileDiff, "/root/", "/localhost:3306/backup/");
     Assert.assertTrue(
         cmd.equals("copy -file /root/test -dest /localhost:3306/backup/test"));
     // Test Copy
     fileDiff.setParameters("-length 1024");
     cmd =
-        CopyScheduler.cmdParsing(fileDiff, "/root/", "/localhost:3306/backup/");
+        LegacyCopyScheduler
+            .cmdParsing(fileDiff, "/root/", "/localhost:3306/backup/");
     Assert.assertTrue(cmd.equals(
         "copy -file /root/test -dest /localhost:3306/backup/test -length 1024"));
     // Test Rename
     fileDiff.setDiffType(FileDiffType.RENAME);
     fileDiff.setParameters("-dest /root/test2 -length 1024");
     cmd =
-        CopyScheduler.cmdParsing(fileDiff, "/root/", "/localhost:3306/backup/");
+        LegacyCopyScheduler
+            .cmdParsing(fileDiff, "/root/", "/localhost:3306/backup/");
     Assert.assertTrue(
         cmd.equals(
             "rename -file /localhost:3306/backup/test -dest /localhost:3306/backup/test2 -length 1024"));
@@ -80,7 +83,7 @@ public class TestCopySchedular extends MiniSmartClusterHarness {
     dfs.mkdirs(new Path(destPath));
     DFSClient client =
         new DFSClient(cluster.getURI(), ssm.getContext().getConf());
-    CopyScheduler copyScheduler = new CopyScheduler(ssm.getContext(),
+    LegacyCopyScheduler legacyCopyScheduler = new LegacyCopyScheduler(ssm.getContext(),
         ssm.getCmdletManager(), client, srcPath, destPath);
     // Write to src
     for (int i = 0; i < 3; i++) {
@@ -90,15 +93,15 @@ public class TestCopySchedular extends MiniSmartClusterHarness {
     Thread.sleep(1500);
     ssm.getMetaStore().deleteAllFileDiff();
     // init forceSync
-    copyScheduler.forceSync(srcPath, destPath);
-    copyScheduler.start();
-    while (copyScheduler.getCachedSize() == 0) {
+    legacyCopyScheduler.forceSync(srcPath, destPath);
+    legacyCopyScheduler.start();
+    while (legacyCopyScheduler.getCachedSize() == 0) {
       Thread.sleep(100);
     }
     while (ssm.getCmdletManager().getActionsSizeInCache() != 0 ||
-        copyScheduler.getCachedSize() != 0) {
+        legacyCopyScheduler.getCachedSize() != 0) {
       System.out.printf("Size of cmdlet %d\n", ssm.getCmdletManager().getActionsSizeInCache());
-      System.out.printf("Size of copy %d\n", copyScheduler.getCachedSize());
+      System.out.printf("Size of copy %d\n", legacyCopyScheduler.getCachedSize());
       System.out.println("In Dead Loop!");
       Thread.sleep(2000);
     }
@@ -107,7 +110,7 @@ public class TestCopySchedular extends MiniSmartClusterHarness {
       Assert.assertTrue(dfs.exists(new Path(destPath + i)));
       System.out.printf("File %d is force copied.\n", i);
     }
-    copyScheduler.stop();
+    legacyCopyScheduler.stop();
   }
 
 
@@ -121,22 +124,22 @@ public class TestCopySchedular extends MiniSmartClusterHarness {
     dfs.mkdirs(new Path(destPath));
     DFSClient client =
         new DFSClient(cluster.getURI(), ssm.getContext().getConf());
-    CopyScheduler copyScheduler = new CopyScheduler(ssm.getContext(),
+    LegacyCopyScheduler legacyCopyScheduler = new LegacyCopyScheduler(ssm.getContext(),
         ssm.getCmdletManager(), client, srcPath, destPath);
-    copyScheduler.start();
+    legacyCopyScheduler.start();
     Thread.sleep(2000);
     for (int i = 0; i < 3; i++) {
       // Write 10 files
       DFSTestUtil.createFile(dfs, new Path(srcPath + i), 1024, (short) 1, 1);
     }
-    while (copyScheduler.getCachedSize() == 0) {
+    while (legacyCopyScheduler.getCachedSize() == 0) {
       Thread.sleep(100);
     }
     while (ssm.getCmdletManager().getActionsSizeInCache() != 0 ||
-        copyScheduler.getCachedSize() != 0) {
+        legacyCopyScheduler.getCachedSize() != 0) {
       Thread.sleep(2000);
       System.out.printf("Size of cmdlet %d\n", ssm.getCmdletManager().getActionsSizeInCache());
-      System.out.printf("Size of copy %d\n", copyScheduler.getCachedSize());
+      System.out.printf("Size of copy %d\n", legacyCopyScheduler.getCachedSize());
       System.out.println("In Dead Loop!");
     }
     for (int i = 0; i < 3; i++) {
@@ -144,7 +147,7 @@ public class TestCopySchedular extends MiniSmartClusterHarness {
       Assert.assertTrue(dfs.exists(new Path(destPath + i)));
       System.out.printf("File %d is copied.\n", i);
     }
-    copyScheduler.stop();
+    legacyCopyScheduler.stop();
   }
 
 
@@ -173,7 +176,7 @@ public class TestCopySchedular extends MiniSmartClusterHarness {
 
     // System.out.println(dfs.getFileStatus(new Path(srcPath + "/" + file1)).getLen());
     List<CopyTargetTask> copyTargetTaskList =
-        CopyScheduler.splitCopyFile(srcPath + "/" + file1,
+        LegacyCopyScheduler.splitCopyFile(srcPath + "/" + file1,
             destPath + "/" + file1, 1,
             FileSystem.get(dfs.getUri(), new Configuration()));
